Java Files
These files are the java files of the various layouts ( pages ) of the app. These files contains the code which stores the information about what to do when an event has occurred such clicking of a button, entering of a text, etc.
The DataBase.java file maintains all the data regarding to the blood bank. i.e. it creates the table, stores information and data in it and gives us the result based on the queries requested. I have used SQLite database for this which is in built database in the Android Studio Itself.